python_streamlit/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.9.0
    # via openai
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
narwhals==1.30.0
    # via altair
numpy==2.2.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==1.66.3
    # via -r requirements.in
packaging==24.2
    # via
    #   altair
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.1.0
    # via streamlit
protobuf==5.29.3
    # via streamlit
pyarrow==19.0.1
    # via streamlit
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.1
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   streamlit
    #   tiktoken
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
streamlit==1.43.2
    # via -r requirements.in
tenacity==9.0.0
    # via streamlit
tiktoken==0.9.0
    # via -r requirements.in
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
